requirements.txt
bcrypt
blinker
certifi
cffi
click
email_validator
Flask
Flask-Bcrypt
Flask-Login
Flask-Mail
Flask-SQLAlchemy
Flask-WTF
itsdangerous
gunicorn
Jinja2
MarkupSafe
mistune
Pillow
pycparser
six
SQLAlchemy
Werkzeug
WTForms
keras==2.2.5
opencv
numpy
tensorflow==1.15.0
